requirements.txt
apiai==1.2.3
speechrecognition==3.8.1
pyaudio==0.2.11
pygame==1.9.6
gtts==2.0.4
watchdog==0.9.0
eel==0.12.0
newsapi-python==0.2.6
pillow==6.2.1
pytesseract==0.3.0
pdf2image==1.10.0
pypdf2==1.26.0
selenium==3.141.0
webdriver-manager==2.5.3
websocket-client==0.48.0
urllib3==1.25.7
requests==2.22.0
python-dotenv==0.11.0
googletrans==2.4.0
pyperclip==1.7.0
playsound==1.2.2
